Corrections

+++ Second_Semester/Tests/FirstTest/findpair.h
@@ -9,7 +9,7 @@ class FindPair
 {
 public:
     FindPair(unsigned int size);
-    short numberIn(unsigned int row, unsigned int column);
+    int numberIn(unsigned int row, unsigned int column);
     short check(unsigned int rowFirst, unsigned int columnFirst, unsigned int rowSecond, unsigned int columnSecond);
     ~FindPair();
 private:
